+++ tools/server-admin/src/intTest/java/org/projectnessie/tools/admin/cli/AbstractContentTests.java
@@ -23,11 +23,22 @@ import static org.projectnessie.versioned.storage.common.objtypes.CommitHeaders.
 import static org.projectnessie.versioned.storage.common.objtypes.ContentValueObj.contentValue;
 import static org.projectnessie.versioned.storage.versionstore.TypeMapping.keyToStoreKey;
 import static org.projectnessie.versioned.storage.versionstore.TypeMapping.objIdToHash;
+import static org.projectnessie.versioned.store.DefaultStoreWorker.payloadForContent;
 
+import com.fasterxml.jackson.databind.JavaType;
+import com.fasterxml.jackson.databind.ObjectMapper;
+import io.quarkus.test.junit.main.LaunchResult;
+import io.quarkus.test.junit.main.QuarkusMainLauncher;
+import java.io.File;
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.List;
 import java.util.Objects;
 import java.util.UUID;
+import org.junit.jupiter.api.io.TempDir;
 import org.projectnessie.model.ContentKey;
+import org.projectnessie.model.IcebergTable;
 import org.projectnessie.model.Namespace;
 import org.projectnessie.nessie.relocated.protobuf.ByteString;
 import org.projectnessie.versioned.Hash;
@@ -40,16 +51,73 @@ import org.projectnessie.versioned.storage.common.persist.Persist;
 import org.projectnessie.versioned.storage.common.persist.Reference;
 import org.projectnessie.versioned.store.DefaultStoreWorker;
 
-abstract class BaseContentPersistTest<OutputType> extends BaseContentTest<OutputType> {
+abstract class AbstractContentTests<OutputType> {
+  private static final ObjectMapper MAPPER = new ObjectMapper();
 
   private final Persist persist;
+  private final Class<OutputType> outputClass;
 
-  BaseContentPersistTest(Persist persist, Class<OutputType> outputClass) {
-    super(outputClass);
+  @TempDir File tempDir;
+
+  protected LaunchResult result;
+  protected List<OutputType> entries;
+
+  protected boolean testNamespaceCreated;
+
+  protected Namespace namespace;
+
+  AbstractContentTests(Persist persist, Class<OutputType> outputClass) {
     this.persist = persist;
+    this.outputClass = outputClass;
   }
 
-  @Override
+  protected void launchNoFile(QuarkusMainLauncher launcher, String... args) {
+    launch(launcher, null, args);
+  }
+
+  protected void launch(QuarkusMainLauncher launcher, String... args) throws Exception {
+    File output = new File(tempDir, "check-content.json");
+    launch(launcher, output, args);
+    JavaType type = MAPPER.getTypeFactory().constructCollectionType(List.class, outputClass);
+    entries = MAPPER.readValue(output, type);
+  }
+
+  private void launch(QuarkusMainLauncher launcher, File outputFile, String... args) {
+    List<String> cmdArgs = new ArrayList<>(Arrays.asList(args));
+
+    if (outputFile != null) {
+      cmdArgs.add("--output");
+      cmdArgs.add(outputFile.getAbsolutePath());
+    }
+
+    result = launcher.launch(cmdArgs.toArray(new String[0]));
+  }
+
+  protected void commit(IcebergTable table) throws Exception {
+    commit(table, true);
+  }
+
+  protected void commit(IcebergTable table, boolean add) throws Exception {
+    ByteString serialized = DefaultStoreWorker.instance().toStoreOnReferenceState(table);
+    commit(
+        ContentKey.of("test_namespace", "table_" + table.getId()),
+        UUID.fromString(Objects.requireNonNull(table.getId())),
+        (byte) payloadForContent(table),
+        serialized,
+        true,
+        add);
+  }
+
+  protected void commit(IcebergTable table, ByteString serialized) throws Exception {
+    commit(
+        ContentKey.of("test_namespace", "table_" + table.getId()),
+        UUID.fromString(Objects.requireNonNull(table.getId())),
+        (byte) payloadForContent(table),
+        serialized,
+        true,
+        true);
+  }
+
   protected void commit(
       ContentKey key,
       UUID contentId,
@@ -96,12 +164,12 @@ abstract class BaseContentPersistTest<OutputType> extends BaseContentTest<Output
     }
 
     CommitLogic commitLogic = commitLogic(persist);
-    CommitObj commit = commitLogic.doCommit(builder.build(), Collections.emptyList());
+    CommitObj commit =
+        Objects.requireNonNull(commitLogic.doCommit(builder.build(), Collections.emptyList()));
 
     referenceLogic(persist).assignReference(refMain, commit.id());
   }
 
-  @Override
   protected Hash getMainHead() {
     Reference main = persist.fetchReference("refs/heads/main");
     return objIdToHash(Objects.requireNonNull(main).pointer());