+++ tools/server-admin/src/intTest/java/org/projectnessie/tools/admin/cli/ITCheckContent.java
@@ -22,21 +22,29 @@ import static org.projectnessie.tools.admin.cli.BaseCommand.EXIT_CODE_CONTENT_ER
 import static org.projectnessie.versioned.store.DefaultStoreWorker.payloadForContent;
 import static org.projectnessie.versioned.testworker.OnRefOnly.onRef;
 
+import io.quarkus.test.junit.TestProfile;
 import io.quarkus.test.junit.main.QuarkusMainLauncher;
 import io.quarkus.test.junit.main.QuarkusMainTest;
 import java.util.UUID;
+import org.junit.jupiter.api.Disabled;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
 import org.projectnessie.model.Content;
 import org.projectnessie.model.ContentKey;
 import org.projectnessie.model.IcebergTable;
 import org.projectnessie.nessie.relocated.protobuf.ByteString;
+import org.projectnessie.quarkus.tests.profiles.BaseConfigProfile;
 import org.projectnessie.versioned.Hash;
+import org.projectnessie.versioned.storage.common.persist.Persist;
 import org.projectnessie.versioned.testworker.OnRefOnly;
 
 @QuarkusMainTest
-abstract class AbstractCheckContentTests {
+@TestProfile(BaseConfigProfile.class)
+@ExtendWith(NessieServerAdminTestExtension.class)
+@Disabled
+class ITCheckContent extends AbstractContentTests<CheckContentEntry> {
 
   private static final UUID CID_1 = UUID.randomUUID();
   private static final UUID CID_2 = UUID.randomUUID();
@@ -52,22 +60,20 @@ abstract class AbstractCheckContentTests {
   private static final IcebergTable table4 =
       IcebergTable.of("meta_444", 4, 5, 6, 7, CID_4.toString());
 
-  private final BaseContentTest<CheckContentEntry> outer;
-
-  AbstractCheckContentTests(BaseContentTest<CheckContentEntry> outer) {
-    this.outer = outer;
+  ITCheckContent(Persist persist) {
+    super(persist, CheckContentEntry.class);
   }
 
   @Test
   public void testEmptyRepo(QuarkusMainLauncher launcher) {
-    outer.launchNoFile(launcher, "check-content");
-    assertThat(outer.result.exitCode()).isEqualTo(0);
+    launchNoFile(launcher, "check-content");
+    assertThat(result.exitCode()).isEqualTo(0);
   }
 
   @Test
   public void testNonExistingKey(QuarkusMainLauncher launcher) throws Exception {
-    outer.launch(launcher, "check-content", "-k", "namespace123", "-k", "unknown12345");
-    assertThat(outer.entries)
+    launch(launcher, "check-content", "-k", "namespace123", "-k", "unknown12345");
+    assertThat(entries)
         .hasSize(1)
         .first()
         .extracting(
@@ -77,13 +83,13 @@ abstract class AbstractCheckContentTests {
             CheckContentEntry::getErrorMessage)
         .containsExactly(
             ContentKey.of("namespace123", "unknown12345"), "ERROR", null, "Missing content");
-    assertThat(outer.result.exitCode()).isEqualTo(EXIT_CODE_CONTENT_ERROR);
+    assertThat(result.exitCode()).isEqualTo(EXIT_CODE_CONTENT_ERROR);
   }
 
   @Test
   public void testParseError(QuarkusMainLauncher launcher) throws Exception {
     ContentKey k1 = ContentKey.of("table123");
-    outer.commit(
+    commit(
         k1,
         UUID.randomUUID(),
         (byte) payloadForContent(Content.Type.ICEBERG_TABLE),
@@ -91,8 +97,8 @@ abstract class AbstractCheckContentTests {
         false,
         true);
 
-    outer.launch(launcher, "check-content");
-    assertThat(outer.entries)
+    launch(launcher, "check-content");
+    assertThat(entries)
         .hasSize(1)
         .first()
         .extracting(
@@ -105,7 +111,7 @@ abstract class AbstractCheckContentTests {
                     && e.getExceptionStackTrace()
                         .contains("Protocol message contained an invalid tag"))
         .containsExactly(ContentKey.of("table123"), "ERROR", null, "Failure parsing data", true);
-    assertThat(outer.result.exitCode()).isEqualTo(EXIT_CODE_CONTENT_ERROR);
+    assertThat(result.exitCode()).isEqualTo(EXIT_CODE_CONTENT_ERROR);
   }
 
   @ParameterizedTest
@@ -113,14 +119,14 @@ abstract class AbstractCheckContentTests {
   public void testWorkerError(int batchSize, QuarkusMainLauncher launcher) throws Exception {
 
     ByteString broken = ByteString.copyFrom(new byte[] {1, 2, 3});
-    outer.commit(table1, broken);
-    outer.commit(table2, broken);
-    outer.commit(table3, broken);
-    outer.commit(table4, broken);
+    commit(table1, broken);
+    commit(table2, broken);
+    commit(table3, broken);
+    commit(table4, broken);
 
     // Note: SimpleStoreWorker will not be able to parse IcebergTable objects
-    outer.launch(launcher, "check-content", "--summary", "--batch=" + batchSize);
-    assertThat(outer.entries.stream())
+    launch(launcher, "check-content", "--summary", "--batch=" + batchSize);
+    assertThat(entries.stream())
         .extracting(
             CheckContentEntry::getKey,
             CheckContentEntry::getStatus,
@@ -136,89 +142,87 @@ abstract class AbstractCheckContentTests {
             tuple(ContentKey.of("test_namespace", "table_" + CID_2), "ERROR", true, true),
             tuple(ContentKey.of("test_namespace", "table_" + CID_3), "ERROR", true, true),
             tuple(ContentKey.of("test_namespace", "table_" + CID_4), "ERROR", true, true));
-    assertThat(outer.result.exitCode()).isEqualTo(EXIT_CODE_CONTENT_ERROR);
-    assertThat(outer.result.getOutputStream())
+    assertThat(result.exitCode()).isEqualTo(EXIT_CODE_CONTENT_ERROR);
+    assertThat(result.getOutputStream())
         .contains(format("Detected %d errors in 5 keys.", batchSize > 1 ? 5 : 4));
   }
 
   @Test
   public void testValidData(QuarkusMainLauncher launcher) throws Exception {
 
-    outer.commit(table1);
-    outer.commit(table2);
+    commit(table1);
+    commit(table2);
 
-    outer.launch(launcher, "check-content", "--show-content");
-    assertThat(outer.entries.stream())
+    launch(launcher, "check-content", "--show-content");
+    assertThat(entries.stream())
         .extracting(
             CheckContentEntry::getKey, CheckContentEntry::getStatus, CheckContentEntry::getContent)
         .containsExactlyInAnyOrder(
-            tuple(ContentKey.of("test_namespace"), "OK", outer.namespace),
+            tuple(ContentKey.of("test_namespace"), "OK", namespace),
             tuple(ContentKey.of("test_namespace", "table_" + CID_1), "OK", table1),
             tuple(ContentKey.of("test_namespace", "table_" + CID_2), "OK", table2));
-    assertThat(outer.result.exitCode()).isEqualTo(0);
+    assertThat(result.exitCode()).isEqualTo(0);
   }
 
   @Test
   public void testValidDataDeletedKey(QuarkusMainLauncher launcher) throws Exception {
 
-    outer.commit(table1); // PUT
-    outer.commit(table1, false); // DELETE
+    commit(table1); // PUT
+    commit(table1, false); // DELETE
 
-    outer.launch(launcher, "check-content", "--show-content");
-    assertThat(outer.entries.stream())
+    launch(launcher, "check-content", "--show-content");
+    assertThat(entries.stream())
         .extracting(
             CheckContentEntry::getKey, CheckContentEntry::getStatus, CheckContentEntry::getContent)
-        .containsExactly(tuple(ContentKey.of("test_namespace"), "OK", outer.namespace));
-    assertThat(outer.result.exitCode()).isEqualTo(0);
+        .containsExactly(tuple(ContentKey.of("test_namespace"), "OK", namespace));
+    assertThat(result.exitCode()).isEqualTo(0);
   }
 
   @Test
   public void testValidDataNoContent(QuarkusMainLauncher launcher) throws Exception {
-    outer.commit(table1);
-    outer.commit(table2);
+    commit(table1);
+    commit(table2);
 
-    outer.launch(launcher, "check-content");
-    assertThat(outer.entries).hasSize(3);
-    assertThat(outer.entries).allSatisfy(e -> assertThat(e.getContent()).isNull());
-    assertThat(outer.result.exitCode()).isEqualTo(0);
+    launch(launcher, "check-content");
+    assertThat(entries).hasSize(3);
+    assertThat(entries).allSatisfy(e -> assertThat(e.getContent()).isNull());
+    assertThat(result.exitCode()).isEqualTo(0);
   }
 
   @Test
   public void testValidDataStdOut(QuarkusMainLauncher launcher) throws Exception {
-    outer.commit(table1);
-    outer.commit(table2);
+    commit(table1);
+    commit(table2);
 
-    outer.launchNoFile(
+    launchNoFile(
         launcher, "check-content", "-s", "--show-content", "--output", "-"); // '-' for STDOUT
-    assertThat(outer.result.getOutputStream())
+    assertThat(result.getOutputStream())
         .anySatisfy(line -> assertThat(line).contains("table_" + CID_1));
-    assertThat(outer.result.getOutputStream())
+    assertThat(result.getOutputStream())
         .anySatisfy(line -> assertThat(line).contains("table_" + CID_2));
-    assertThat(outer.result.getOutputStream())
-        .anySatisfy(line -> assertThat(line).contains("meta_111"));
-    assertThat(outer.result.getOutputStream())
-        .anySatisfy(line -> assertThat(line).contains("meta_222"));
-    assertThat(outer.result.exitCode()).isEqualTo(0);
-    assertThat(outer.result.getOutputStream()).contains("Detected 0 errors in 3 keys.");
+    assertThat(result.getOutputStream()).anySatisfy(line -> assertThat(line).contains("meta_111"));
+    assertThat(result.getOutputStream()).anySatisfy(line -> assertThat(line).contains("meta_222"));
+    assertThat(result.exitCode()).isEqualTo(0);
+    assertThat(result.getOutputStream()).contains("Detected 0 errors in 3 keys.");
   }
 
   @Test
   public void testErrorOnly(QuarkusMainLauncher launcher) throws Exception {
-    outer.commit(table1);
-    outer.launch(launcher, "check-content", "--error-only");
-    assertThat(outer.entries).hasSize(0);
-    assertThat(outer.result.exitCode()).isEqualTo(0);
+    commit(table1);
+    launch(launcher, "check-content", "--error-only");
+    assertThat(entries).hasSize(0);
+    assertThat(result.exitCode()).isEqualTo(0);
   }
 
   @Test
   public void testHashWithBrokenCommit(QuarkusMainLauncher launcher) throws Exception {
-    outer.commit(table1);
+    commit(table1);
 
-    Hash hash = outer.getMainHead();
+    Hash hash = getMainHead();
 
     UUID contentId = UUID.randomUUID();
     OnRefOnly val = onRef("123", contentId.toString());
-    outer.commit(
+    commit(
         ContentKey.of("test_namespace", "table_" + val.getId()),
         contentId,
         (byte) payloadForContent(val),
@@ -226,12 +230,12 @@ abstract class AbstractCheckContentTests {
         true,
         true);
 
-    outer.launch(launcher, "check-content", "--hash", hash.asString());
-    assertThat(outer.entries.stream())
+    launch(launcher, "check-content", "--hash", hash.asString());
+    assertThat(entries.stream())
         .extracting(CheckContentEntry::getKey, CheckContentEntry::getStatus)
         .containsExactlyInAnyOrder(
             tuple(ContentKey.of("test_namespace"), "OK"),
             tuple(ContentKey.of("test_namespace", "table_" + CID_1), "OK"));
-    assertThat(outer.result.exitCode()).isEqualTo(0);
+    assertThat(result.exitCode()).isEqualTo(0);
   }
 }